+++ src/main/java/com/ingemark/webshopbasics/service/OrderItemService.java
@@ -1,6 +1,8 @@
 package com.ingemark.webshopbasics.service;
 
+import java.math.BigDecimal;
+
 public interface OrderItemService {
 
-	Double selectTotalByOrderId(Long pIdOrder);
+	BigDecimal selectTotalByOrderId(Long pIdOrder);
 }
